requirements.txt
altair==5.5.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
databricks-sdk==0.43.0
databricks-sql-connector==4.0.0
et_xmlfile==2.0.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
idna==3.10
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lz4==4.4.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.25.2
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pinecone-client==3.0.1
pillow==11.1.0
protobuf==5.29.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
six==1.17.0
smmap==5.0.2
streamlit==1.42.0
tenacity==9.0.0
thrift==0.20.0
toml==0.10.2
tornado==6.4.2
torch==2.6.0
transformers==4.48.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
